refactor: 지도 탭 컴포넌트 분리
